+++ rakaia/utils/alert.py
@@ -149,7 +149,8 @@ class ToolTips(BaseModel):
                       "default-scaling": "Reset the lower bound to 0, and the upper bound to the 99th percentile",
                       "region-annotate-all": "By default, only the most recently drawn shape is annotated.",
                       "query-obj-min": "Define a minimum number of objects required per ROI when querying from the UMAP. "
-                                       "ROIs without the minimum object number will not be displayed in the gallery."}
+                                       "ROIs without the minimum object number will not be displayed in the gallery.",
+                      "download-region-statistics": "Download the region statistic table (CSV)"}
 
 
 class PanelMismatchError(Exception):